Remove unused current_case
