+++ src/ert/gui/simulation/ensemble_experiment_panel.py
@@ -6,7 +6,6 @@ from qtpy.QtWidgets import QFormLayout, QLabel
 from ert.gui.ertnotifier import ErtNotifier
 from ert.gui.ertwidgets import StringBox, TextModel
 from ert.gui.ertwidgets.copyablelabel import CopyableLabel
-from ert.gui.ertwidgets.ensembleselector import EnsembleSelector
 from ert.gui.ertwidgets.models.activerealizationsmodel import ActiveRealizationsModel
 from ert.mode_definitions import ENSEMBLE_EXPERIMENT_MODE
 from ert.run_models import EnsembleExperiment
@@ -54,8 +53,6 @@ class EnsembleExperimentPanel(SimulationConfigPanel):
 
         layout.addRow("Ensemble name:", self._ensemble_name_field)
 
-        self._ensemble_selector = EnsembleSelector(notifier)
-        layout.addRow("Current ensemble:", self._ensemble_selector)
         runpath_label = CopyableLabel(text=run_path)
         layout.addRow("Runpath:", runpath_label)
