+++ src/ert/gui/simulation/multiple_data_assimilation_panel.py
@@ -204,3 +204,11 @@ class MultipleDataAssimilationPanel(SimulationConfigPanel):
 
     def setWeights(self, weights):
         self.weights = str(weights)
+
+    def _realizations_from_fs(self):
+        ensemble = str(self._ensemble_selector.currentText())
+        if ensemble:
+            mask = self.notifier.storage.get_ensemble_by_name(
+                ensemble
+            ).get_realization_mask_with_parameters()
+            self._active_realizations_field.model.setValueFromMask(mask)
