+++ src/ert/gui/simulation/simulation_config_panel.py
@@ -22,11 +22,3 @@ class SimulationConfigPanel(QWidget):
     @staticmethod
     def getSimulationArguments() -> Dict[str, Any]:
         return {}
-
-    def _realizations_from_fs(self):
-        ensemble = str(self._ensemble_selector.currentText())
-        if ensemble:
-            mask = self.notifier.storage.get_ensemble_by_name(
-                ensemble
-            ).get_realization_mask_with_parameters()
-            self._active_realizations_field.model.setValueFromMask(mask)
